requirements.txt
fastapi[all]==0.110.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
